refactor: move helper to openidStrategy.js
